+++ include/MinLenghtProblem.h
@@ -36,8 +36,8 @@ public:
     void set_graph(Graph G);
     void Generate_graph();
     void Generate_XSortedGraph();
-    float SimpleSolution(int node_pair[2]);
-    float DCSolution();
+    int SimpleSolution(int node_pair[2]);
+    int DCSolution();
 };
 
 #endif // MINLENGHTPROBLEM_H
